% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - gene-match-o-matic 2000.R
\name{calc_genetic_similarity}
\alias{calc_genetic_similarity}
\title{Calculate the genetic similarity between two individuals (based on SNPs)}
\usage{
calc_genetic_similarity(sample1, sample2, cutoff_NA_collective = 50)
}
\arguments{
\item{SNPs1}{the first vector of SNPs}

\item{SNPs2}{the second vector of SNPs}
}
\description{
takes two vectors of SNPs,
Each element in the vector should be the genotypes of one SNP, written as bases, e.g "CT" "CC" "AT" etc. Order, e.g. "CT" or "TC" is not taken into account.
It's important that the order of SNPs in the two vectors are the same.
Returns an integer between 1 and 0 (0 = 0% match, 1=100% match)
}
\keyword{internal}
