% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - fishTools.R
\name{calc_TGC_df}
\alias{calc_TGC_df}
\title{Calculates Thermal Growth Coefficient TGC}
\usage{
calc_TGC_df(df_fish, df_temp, period)
}
\description{
Calculates Thermal Growth Coefficient TGC
}
\keyword{internal}
