% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - base tools.R
\name{check_columns}
\alias{check_columns}
\title{Cecks if certain columns exist in a dataset and returns an error message if not}
\usage{
check_columns(dataset, columns, preMessage = "Missing columns:")
}
\description{
Cecks if certain columns exist in a dataset and returns an error message if not
}
\keyword{internal}
