% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - base tools.R
\name{checkup}
\alias{checkup}
\title{Applies a function on the column of a dataframe and then returns that dataframe}
\usage{
checkup()
}
\arguments{
\item{df}{A dataframe}

\item{column}{The name of the column (string) that we want apply the function to}

\item{fun}{The function we use on the column}
}
\description{
Applies a function on the column of a dataframe and then returns that dataframe
}
