% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - base genotools.R
\name{clean_ID_df}
\alias{clean_ID_df}
\title{In a dataframe, converts messy names and ID's to tidy clean ones.}
\usage{
clean_ID_df(
  df,
  column_name = "ID",
  identifier = "",
  identifier_left = F,
  numLength = F,
  prefix = "",
  na_remove = T,
  keep_name = F,
  numeric = F
)
}
\arguments{
\item{df}{The data frame}

\item{identifier}{ID's need to be formated with a number and following identifier, e.g "34_individuals2019" where "_individuals2019" is the identifier. Any entries not matching this format will be removed.}

\item{identifier_left}{Wether the identifier is on the left hand (T) or right-hand (R) side of the number}

\item{numLength}{if you want leading zeroes, use this parameter to specify the length of the number, e.g "8" for 00000342}

\item{prefix}{if you want a prefix in the new cleaned ID. Ex: "individuals2019_" will give you "individuals2019_0034"}

\item{na_remove}{if you want to remove any rows that don't follow your pattern (otherwise, they'll turn to NA). Default is True.}

\item{column}{The name of a column containing dirty IDs}
}
\description{
For sorting out column with long and complicated identifiers or row names, where the true ID of a row is hidden in a string.\cr
E.g: Make "dirty" ID's like "A0006_3911_BT-F1_GTCGTCTA_run20190930N" turn into "clean" ID's like 3991_BT
}
