% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - gene-match-o-matic 2000.R
\name{count_matches}
\alias{count_matches}
\title{subfunction of calc_genetic_similarity}
\usage{
count_matches(snp1, snp2)
}
\description{
subfunction of calc_genetic_similarity
}
\keyword{internal}
