% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - base genotools.R
\name{determineSex}
\alias{determineSex}
\title{Converting sdy to F or M}
\usage{
determineSex(dataframe, column, cutoff)
}
\description{
Used on dataframes, for determining sex based on SDY in a given column
}
