% Generated by roxygen2 (4.0.1): do not edit by hand
\name{display.colors}
\alias{display.colors}
\title{Display a vector of colors for viewing}
\usage{
display.colors(cols)
}
\arguments{
\item{cols}{A character vector of colors}
}
\value{
An tiled image with color labels as a \code{gg} object
}
\description{
Display a vector of colors for viewing
}
\examples{
display.colors(colors()[58:61])

m <- matrix(c(1,0.5, 0, 0.5, 0.2, 0.9, 0.2, 1, 1, 0.3, 0.6, 0.2), ncol = 3)
display.colors(rgb(m))

pal <- colorRampPalette(c('blue', 'yellow', 'pink'))
display.colors(pal(10))
}

