% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - family tools.R
\name{find_familyIDs}
\alias{find_familyIDs}
\title{Obtains family IDs based on known mothers and fathers}
\usage{
find_familyIDs(df, df_families)
}
\description{
Works on a dataframe of individuals (rows) including two columns ID_ma and ID_pa, which are the ID's of that individual's parents.
\cr
Then looks up a dataframe of family info, df_families, with rows as families, and finds the ID's of each individual's family
}
