% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - family tools.R
\name{find_familyInfo}
\alias{find_familyInfo}
\title{Obtains info about a given family (mother x father) and adds that info to their offspring}
\usage{
find_familyInfo(df, df_families, columns)
}
\description{
Works on a dataframe of individuals (rows) including two columns ID_ma and ID_pa, which are the ID's of that individual's parents.
For each individual, looks up another dataset of family info to attach family info to that individual (based on ID of its mother and father
df_families structure: must include columns ID_ma and ID_pa for each row (family).
}
