% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - base tools.R
\name{funkyTranspose}
\alias{funkyTranspose}
\title{funkyTranspose
https://stackoverflow.com/questions/6645524/what-is-the-best-way-to-transpose-a-data-frame-in-r-and-to-set-one-of-the-column
Credits: mortonjt and nzcoops}
\usage{
funkyTranspose(df)
}
\description{
funkyTranspose
https://stackoverflow.com/questions/6645524/what-is-the-best-way-to-transpose-a-data-frame-in-r-and-to-set-one-of-the-column
Credits: mortonjt and nzcoops
}
\keyword{internal}
