% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation - genotools.R
\docType{package}
\name{Genotools}
\alias{Genotools}
\title{Genotools: A collection of various tools for working with large salmon datasets (individuals and genotypes)}
\description{
Genotools: A collection of various tools for working with large salmon datasets (individuals and genotypes)
}
