% Generated by roxygen2 (4.0.1): do not edit by hand
\name{in_interval}
\alias{in_interval}
\title{check if values are within a given interval}
\usage{
in_interval(x, low, high, low.closed = T, high.closed = T)
}
\arguments{
\item{x}{a numeric vector}

\item{low,high}{the limits of the reference interval}

\item{low.closed,high.closed}{set to TRUE for inclusion of the corresponding limit as within the interval. e.g. to test an open interval (x,y), set both to FALSE; to test a closed interval [x,y], set both to TRUE}
}
\value{
vector of logicals
}
\description{
check if values are within a given interval
}
\examples{
vec <- 1:10
in_interval(vec, 2, 8, T, T)
in_interval(vec, 2, 8, T, F)
in_interval(vec, 2, 8, F, F)
}

