% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - base tools.R
\name{lookup}
\alias{lookup}
\title{For looking up variables from one dataset and then add them to another one.}
\usage{
lookup(
  df_samples,
  df_lookup,
  id_column,
  value_column,
  default = NA,
  overwrite = T
)
}
\arguments{
\item{df_samples}{samples to look up}

\item{df_lookup}{dataframe to look up against}

\item{id_column}{common column between the two sets containing unique identifiers for rows}

\item{value}{the value that is looked up and added to df_samples}
}
\description{
Use to add a column (value) to a dataset (samples), from another dataset (lookup), based on an identifier that exists in both (id_lookup)
}
