% Generated by roxygen2 (4.0.1): do not edit by hand
\name{my.gc}
\alias{my.gc}
\title{Collect all garbage, i.e. clear temporary memory. Keeps running /code{gc()} until output
 doesn't change}
\usage{
my.gc()
}
\description{
Collect all garbage, i.e. clear temporary memory. Keeps running /code{gc()} until output
 doesn't change
}

