% Generated by roxygen2 (4.0.1): do not edit by hand
\name{my.grepl}
\alias{my.grepl}
\title{A list version of grepl}
\usage{
my.grepl(vec, ..., fun = c("any", "all"))
}
\arguments{
\item{vec}{character vector of regular expressions}

\item{...}{arguments passed to /code{grepl}}

\item{fun}{specifies the aggregation of logicals to /code{vec}. If /code{'any'} is #' specified, /code{TRUE} will return if any of the regular expressions are found.
If /code{'all'} is specified, /code{TRUE} will return only if each regular
 expression is found.}
}
\value{
logical vector
}
\description{
This function extends grepl to searche a vector of regular expressions, to ask:
"which elements contain any/all of these regular expressions?"
}
\examples{
words <- c('antihistamine', 'antechamber', 'octopus')
re <- c('anti', 'am')
my.grepl(re, words, fun = 'any')
my.grepl(re, words, fun = 'all')
}

