% Generated by roxygen2 (4.0.1): do not edit by hand
\name{peek}
\alias{peek}
\title{Get a random sample from a data object}
\usage{
peek(x, n = 6L)
}
\arguments{
\item{x}{a vector, matrix, or data.frame}

\item{n}{an integer specifying the number of elements (vector) or rows
(matrix/data.frame) to display}
}
\value{
a random subset of x. Elements will be in the same order as in x, and
  output is of the same class as x.
}
\description{
Analogous to \code{head} and \code{tail}, but takes a random sample of rows
to display and prints them in order
}
\examples{
\dontrun{
set.seed(123);
 peek(USArrests)
 }
}

