% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - fishTools.R
\name{predict_weights_df}
\alias{predict_weights_df}
\title{predict_weights_df}
\usage{
predict_weights_df(
  df_fish,
  df_temp,
  dates,
  startPeriod = 999,
  tgc_p_modifier = 1,
  temp_abs_modifier = 0
)
}
\description{
predict_weights_df
}
\keyword{internal}
