% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - ransamp.R
\name{ransamp}
\alias{ransamp}
\title{Random, condition based samplig of a set of individuals}
\usage{
ransamp(
  df_f,
  ofEach,
  nOfEach,
  noShareWithin,
  priBy,
  usedupl = F,
  checkExternal = F,
  identifier = ""
)
}
\arguments{
\item{df_f}{The dataframe to work on}

\item{ofEach}{The parameters we want to use, as a vector of column names, e.g: ofEach=c("sex","age","tank")}

\item{nOfEach}{How many individuals within each combination of the "ofEach" parameters that should be selected}

\item{noShareWithin}{one or more sets of paremeters, where unique combinations can't be shared. For example, if we want no more than a single individual from any given family, within a tank, a set would be c("family","tank"), meaning that no individuals can have the same of both family and tank. Supplied within a list, so: list(c("family","tank")). Add more sets within the list if needed, for example: list(c("tank","father"),c("tank","mother)))}

\item{priBy}{If some individuals are to be prioritized over others, specify the name of the column containing prioritization info. This must ba number, and lower numbers are prioritized. (E.g, individuals with "1" are prioritized over individuals iwth "2")}

\item{checkExternal}{...}

\item{identifier}{...}

\item{useDupl}{If all individuals that are selected within each combination may be be used, or just one of them.}
}
\description{
Takes a dataframe of individuals (rows) with certain attributes (columns), and samples a given number of individuals from all possible combinations of a given set of paramters
For example, can sample one individual of ech combination of sex and age
}
