% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - base genotools.R
\name{renameGenotype}
\alias{renameGenotype}
\title{renameGenotype}
\usage{
renameGenotype(
  dataframe,
  column,
  LUT = c(`1` = "1 1", `2` = "1 2", `3` = "2 2")
)
}
\description{
renameGenotype
}
\keyword{internal}
