% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - gene-match-o-matic 2000.R
\name{resolve_conflicts}
\alias{resolve_conflicts}
\title{resolve_conflicts}
\usage{
resolve_conflicts(df_samples, df_similarities)
}
\description{
subfunction of assign_closest_matches
}
\keyword{internal}
