% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - base tools.R
\name{selRandom}
\alias{selRandom}
\title{Selects random rows from a dataframe}
\usage{
selRandom(df, n)
}
\description{
Takes a dataframe and a number n
Returns return n randomly selected rows from the dataframe (as a dataframe)
}
