% Generated by roxygen2 (4.0.1): do not edit by hand
\name{swap}
\alias{swap}
\title{Replace values in a vector}
\usage{
swap(vec, from, to)
}
\arguments{
\item{vec}{a vector}

\item{from}{a vector of values that will be replaced in \code{vec}, if present}

\item{to}{a vector of the same length as \code{from}, with replacement values for each
corresponding element in the same index}
}
\value{
a vector of the same length as \code{vec}, with values replaced
}
\description{
Replace values in a vector
}
\examples{
x <- LETTERS[1:10]; print(x)
swap(x, c('J' 'K', 'B'), c('j', 'k', 'b'))
}

